Removing duplicate test code: check_interhost_subhalos done.
